+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cli: Added support for new referral code management instructions.
 - cli: Added the `other close-idl` command for closing IDL accounts.
 - cli: Added the `other resize-idl` command for resizing IDL accounts.
+- cli: Added the `other set-idl-authority` command.
 - tests: Added an integration testing suite `integration_test` to the `gmsol` tests.
 - docs: Created a `CHANGELOG.md` file to document project updates.